// models/store.d.ts
export interface RootStoreState extends StoreState {
    accidents: AccidentsState;
    Accounting: {
        assetsDetails: AssetsDetailsState;
        assetTransactions: AssetTransactionsDetailsState;
        breakdownBrokerPending: BreakdownBrokerPaymentsDetailState;
        brokerMetricsDetails: BrokerMetricsDetailsState;
        brokerReviewsDetails: BrokerReviewsDetailsState;
        brokers: BrokersDetailsState;
        contactsDetails: ContactsDetailsState;
        drivers: DriversDetailsState;
        factoring: FactoringDetailsState;
        fuelReport: FuelReportsState;
        generateInvoice: GenerateInvoiceDetailsState;
        invoiceBoard: InvoiceBoardDetailsState;
        invoices: InvoicesDetailsState;
        mileageReport: MileageReportsState;
        operationalPayments: OperationalTransactionsDetailsState;
        payablePaymentsSummary: PayablePaymentSummaryState;
        payableSumary: PayableSummaryState;
        payrollFinancialSummary: PayrollFinancialSummaryState;
        payrollMetricsSummary: PayrollMetricsSummaryState;
        recievables: ReceivablesSummaryState;
        receivedPayments: ReceivedPaymentDetailsState;
        taskDetails: TaskDetailsState;
        tollReport: TollReportsState;
        vendorPayments: VendorPaymentsDetailsState;
        vendors: VendorsDetailsState;
    }
    alerts: AlertState;
    applicants: ApplicantsState;
    Broker: {
        brokersDetails: BrokersDetailsState;
        brokerSummary: BrokerSummaryState;
        carrierSalesLeadersDetails: CarrierSalesLeadersState;
        carriersDetails: CarriersDetailsState;
        contactsDetails: ContactsDetailsState;
        contractRates: ContractRatesDetailsState;
        customerSalesLeadersDetails: CustomerSalesLeadersState;
        customersDetails: CustomersDetailsState;
        docummentsDetails: DocumentsDetailsState;
        invoicingDetails: InvoicingDetailsState;
        loadHistory: LoadHistoryDetailsState;
        loadsDetails: LoadsDetailsState;
        notesDetails: NotesDetailsState;
        openLoads: OpenLoadsDetailsState;
        quotes: QuotesDetailsState;
        shipmentsDetails: ShipmentsDetailsState;
        topCarriersDetails: TopCarriersState;
        topCustomers: TopCustomersDetailsState;
        yearlySummary: YearlySummaryState;
    }
    brokers: BrokersDetailsState;
    carriers: CarriersDetailsState;
    client: ClientState;
    common: CommonState;
    company: CompanyState;
    devices: DevicesDetailsState;
    dispatch: DispatchSummaryState;
    documents: DocumentsDetailsState;
    driver: DriverState;
    driverBoard: DriversDetailsState;
    drugTest: DrugTestState;
    grossBoard: GrossBoardDetailsState;
    grossFinancial: GrossFinancialSummaryState;
    incidentRegistry: IncidentRegistryDetailsState;
    inspectionReport: InspectionReportDetailsState;
    inspections: InspectionsDetailsState;
    loads: LoadDetailsState;
    locations: LocationsDetailsState;
    onboard: OnboardState;
    permits: PermitsDetailsState;
    safetyMetrics: SafetyMetricsSummaryState;
    services: ServicesDetailsState;
    stops: StopState;
    trailer: TrailerState;
    training: TrainingState;
    truck: TruckState;
    vehicles: VehiclesState;
    violationsDetails: ViolationsState;
    violationsPoints: ViolationsState;
}